feat: Update Prisma schema to include latitude and longitude for sitters, rename pets field in user model, and adjust import paths for Prisma in API files
