+++ prisma/schema.prisma
@@ -105,6 +105,8 @@ model sitter {
   approval_status_id      Int                       @default(1)
   status_updated_at       DateTime?                 @db.Timestamp(6)
   admin_note              String?                   @db.VarChar(1000)
+  latitude                Float?
+  longitude               Float?
   review                  review[]
   sitter_approval_status  sitter_approval_status    @relation(fields: [approval_status_id], references: [id], onDelete: NoAction, onUpdate: NoAction)
   user                    user                      @relation(fields: [user_sitter_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
@@ -171,7 +173,7 @@ model user {
   status_updated_at       DateTime?                 @db.Timestamp(6)
   admin_note              String?                   @db.VarChar(1000)
   admin                   admin?
-  pet                     pet[]
+  pets                    pet[]
   report                  report[]
   review                  review[]
   sitter                  sitter[]
@@ -191,8 +193,6 @@ model user_role {
   role_id Int
   role    role @relation(fields: [role_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
   user    user @relation(fields: [user_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
-
-  @@unique([user_id, role_id])
 
   @@unique([user_id, role_id])
 }
